fix: commit working state
