+++ .env
@@ -1,6 +1,7 @@
 REACT_APP_SUPPORTED_ETHEREUM_NETWORKS="polygon,mumbai,mainnet,kovan,fuji,avalanche"
 REACT_APP_DEFAULT_ETHEREUM_NETWORK="mainnet"
 REACT_APP_ENABLE_CACHING_BACKEND="true"
+REACT_APP_ENABLE_TESTNET="true"
 REACT_APP_RATES_HISTORY_ENDPOINT="https://aave-api-v2.aave.com/data/rates-history"
 REACT_APP_ENABLE_NASH="true"
 # REACT_APP_FORTMATIC_KEY_MAINNET=<FORTMATIC_KEY_MAINNET>
